Like usual, I'm just hacking things inside of other methods to experiment quickly.
Current experiment shows that it is actually pretty easy to manage the current playlist (I successfully added a new track from the library).

From here, I need to give at least a remove index for the playlist and then search. I'm thinking that it should just enumerate the track list and then drop everything in Lucene and search that. All we need to store is the mediaId and MediaType